feat: move classes from main to own files
